+++ src/main/java/com/sparta/cream/dto/response/CreatePaymentResponse.java
@@ -1,5 +1,7 @@
 package com.sparta.cream.dto.response;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 
 @Getter
@@ -8,12 +10,12 @@ public class CreatePaymentResponse {
 	private final String paymentId;
 	private final String status;
 	private final String productName;
-	private final Long amount;
+	private final BigDecimal amount;
 	private final String email;
 	private final String customerName;
 	private final String customerPhoneNumber;
 
-	public CreatePaymentResponse(Long id, String paymentId, String status, String productName, Long amount, String email, String customerName, String customerPhoneNumber) {
+	public CreatePaymentResponse(Long id, String paymentId, String status, String productName, BigDecimal amount, String email, String customerName, String customerPhoneNumber) {
 		this.id = id;
 		this.paymentId = paymentId;
 		this.status = status;
