+++ src/main/java/com/sparta/cream/dto/response/RefundPaymentResponse.java
@@ -1,14 +1,16 @@
 package com.sparta.cream.dto.response;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 
 @Getter
 public class RefundPaymentResponse {
     private final Long refundId;
-    private final Long cancelledAmount;
+    private final BigDecimal cancelledAmount;
     private final String status;
 
-	public RefundPaymentResponse(Long refundId, Long cancelledAmount, String status) {
+	public RefundPaymentResponse(Long refundId, BigDecimal cancelledAmount, String status) {
 		this.refundId = refundId;
 		this.cancelledAmount = cancelledAmount;
 		this.status = status;
